fix: prevent duplicate follow-up messages
